Styling on StudentInterview
