+++ src/components/App/App.js
@@ -3,13 +3,13 @@ import HomePage from '../HomePage/HomePage';
 import NavBar from '../NavBar/NavBar';
 import LoginForm from '../LoginForm/LoginForm';
 import Dashboard from '../Dashboard/Dashboard';
-import StudentInterviewContainer from '../StudentInterviewContainer/StudentInterviewContainer';
 import StudentSearch from '../StudentSearch/StudentSearch';
 import AdminQuestions from '../AdminQuestions/AdminQuestions';
 import { InterviewsContext, UserContext, QuestionsContext } from '../../Context';
 import { Route, Redirect } from 'react-router-dom';
 import Interview from '../Interview/Interview';
 import StudentQuestions from '../StudentQuestions/StudentQuestions';
+import StudentInterviews from '../StudentInterviews/StudentInterviews';
 
 // Apollo
 import ApolloClient from 'apollo-boost';
@@ -39,7 +39,7 @@ export const App = () => {
               <Route exact path='/'><HomePage /></Route>
               <Route exact path='/login' render={() => user ? (<Redirect to='/dashboard'/>) : <LoginForm />}/>
               <Route exact path='/dashboard' render={() => !user ? (<Redirect to='/login'/>) : <Dashboard /> } />
-              <Route exact path='/student-interviews' render={() => !user ? (<Redirect to='/login'/>) : <StudentInterviewContainer /> } />
+              <Route exact path='/student-interviews' render={() => !user ? (<Redirect to='/login'/>) : <StudentInterviews /> } />
               <Route exact path='/student-questions' render={() => !user ? (<Redirect to='/login'/>) : <StudentQuestions /> } />
               <Route exact path='/interview' render={() => !user.currentInterview ? (<Redirect to='/login'/>) : <Interview /> } />
               <Route exact path='/students' render={() => user.role !== 2 ? (<Redirect to='/login'/>) : <StudentSearch />} />
