+++ src/components/InterviewQuestion/InterviewQuestion.js
@@ -26,7 +26,7 @@ export const Question = props => {
   }
 
   return (
-    <div className={'interview-card shadow ' + whereAmI(props.cur)}>
+    <div className={'interview-question-card shadow ' + whereAmI(props.cur)}>
       <h1 className='interview-card-question'>Q: {props.question}</h1>
       <h1 className='interview-card-position'>{props.pos}</h1>
