+++ src/components/StudentInterview/StudentInterview.js
@@ -1,13 +1,14 @@
 import React from 'react';
 
-export const StudentInterview = ({ summary, id, collapseModal, identifyInterview, collapsed }) => {
+export const StudentInterview = ({ summary, id, showDetails, detailed }) => {
 
+  const isDetailed = () => {if (detailed) {return ' shown'} else {return ''}}
   return (
-    <section className='student-interview'>
-      <p>{summary}</p>
-      <button
-        onClick={() => {collapseModal(!collapsed); identifyInterview(id)}}
-      >Expand</button>
+    <section className='interview-card' onClick={e => showDetails(id)}>
+      <h3>{id}</h3>
+      <div className={'details' + isDetailed()}>
+        {summary}
+      </div>
     </section>
   )
 }
