--- src/components/StudentInterviewModal/StudentInterviewModal.js
@@ -1,12 +1,0 @@
-import React from 'react';
-
-export const StudentInterviewModal = ({ id, summary }) => {
-  return (
-    <>
-    <div>{id}</div>
-    <p>{summary}</p>
-    </>
-  )
-}
-
-export default StudentInterviewModal;
