+++ src/styles/_InterviewQuestion.scss
@@ -6,7 +6,7 @@
   box-sizing: border-box;
 }
 
-.interview-card {
+.interview-question-card {
   background: $card-bg;
   color: $card-fg;
   padding: 2rem;
